gitflow-feature-stash: melhora_da_arquitetura
Modularização das views e reorganização dos templates.
