--- pessoas/admin.py
@@ -1,13 +1,0 @@
-from django.contrib import admin
-from .models import Pessoa
-
-
-class ListandoPessoas(admin.ModelAdmin):
-    list_display = ('id', 'nome')
-    list_display_links = ('id', 'nome')
-    search_fields = ('nome',)
-    list_per_page = 10
-
-
-
-admin.site.register(Pessoa, ListandoPessoas)
