+++ receitas/views/__init__.py
@@ -1,0 +1,2 @@
+from .receita import *
+from .busca import *
