+++ usuarios/urls.py
@@ -7,9 +7,5 @@ urlpatterns = [
     path('login', views.login, name='login'),
     path('dashboard', views.dashboard, name='dashboard'),
     path('logout', views.logout, name='logout'),
-    path('cria/receita', views.cria_receita, name='cria_receita'),
-    path('deleta/<int:receita_id>', views.deleta_receita, name='deleta_receita'),
-    path('edita/<int:receita_id>', views.edita_receita, name='edita_receita'),
-    path('atualiza_receita', views.atualiza_receita, name='atualiza_receita'),
 
 ]
